+++ src/commands/index.js
@@ -1,0 +1,3 @@
+module.exports = {
+    schemaToJson: require('./schemaToJson')
+};