+++ src/helpers/index.js
@@ -1,3 +1,10 @@
 module.exports = {
-    parsePgType: require('./parsePgType')
+    parsePgType: require('./parsePgType'),
+
+    /* readers */
+    readJsonTable: require('./readJsonTable'),
+
+    /* writers */
+    toCreateTable: require('./toCreateTable'),
+    toMarkdown: require('./toMarkdown')
 };